requirements.txt
requests>=2.25.0
beautifulsoup4>=4.9.0
cssutils>=2.3.0
Pillow>=8.0.0
colorthief>=0.2.1
PyYAML>=6.0
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=0.900